Rename src/ncomp_internal directory to ncomp_build
This should make it less ambiguous that these headers are intended only
for builds and not for installation.
