+++ src/ncomp_build/util.hpp
